+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 #### Updates
 
 * Update to DJI SDK 3.1.1
-* Update to latest Video Previewer (include as framework instead of code)
+* Update to latest Video Previewer
 * Use hardware decoding for Video Previewer when the connected model supports it
 
 ### [1.4.2b3](https://github.com/dbaldwin/DronePan/releases/tag/1.4.2b3)